refactor: create bucket file reader library
